Migrated project to vscode environment from visual studio
Visual Studio felt cluttered, too complicated for the project structure. I had to spend some time with setting up the C++ environment in vscode but it was worth it in the end and proved to be a great lesson on g++ and mingw.

Adding external libraries to intellisense and making sure the compiler handles them was the most challenging part, knowledge will prove useful in later projects.